% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_util.R
\name{add_irrelevant_features}
\alias{add_irrelevant_features}
\title{A function to add irrelevant features to existing data}
\usage{
add_irrelevant_features(df0, p, distn, ...)
}
\arguments{
\item{df0}{dataframe; the data}

\item{p}{integer; number of irrelevant covariates to be added}

\item{distn}{One of 'gaussian', 'student-t', 'pareto', 'uniform', 'discrete_uniform', 'poisson', 'gamma'.}

\item{...}{Optional parameters to be passed to "generate_features".}
}
\description{
A function to add irrelevant features to existing data
}
\details{
This function calls 'generate_features'.
See '?generate_features' for details.
}
