% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/artificialData.R
\docType{package}
\name{artificialData}
\alias{artificialData}
\alias{artificialData-package}
\title{A package for simulating data}
\description{
This package simulates artifical data following a GLM model with dependent / independent, numerical / ordinal covariates.
}
\author{
Jackson Kwok
}

