% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_noise_ratio.R
\name{atomic_find_signal_label}
\alias{atomic_find_signal_label}
\title{Categorise signal-noise ratio}
\usage{
atomic_find_signal_label(signal_noise_ratio)
}
\description{
Categorise signal-noise ratio
}

