% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_search.R
\name{block_seq_by_num}
\alias{block_seq_by_num}
\title{[Core] Partition a sequence into blocks given number of blocks}
\usage{
block_seq_by_num(dim_param, block_num)
}
\description{
[Core] Partition a sequence into blocks given number of blocks
}

