% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stochastic_search.R
\name{block_seq_by_size}
\alias{block_seq_by_size}
\title{[Core] Partition a sequence into blocks given block size}
\usage{
block_seq_by_size(dim_param, block_size)
}
\description{
[Core] Partition a sequence into blocks given block size
}

