% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_noise_ratio.R
\name{create_labels_table}
\alias{create_labels_table}
\title{Create a dataframe mapping labels to intervals.}
\usage{
create_labels_table(labels = c("S", "M", "L"), partition = c(0, exp(c(0.5,
  1.5)), Inf))
}
\arguments{
\item{labels}{vector; the labels.}

\item{partition}{vector; the breakpoint partitioning the positive real line.}
}
\description{
Create a dataframe mapping labels to intervals.
}

