% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_noise_ratio.R
\name{diff_ratio}
\alias{diff_ratio}
\title{Compare the signal-noise ratios}
\usage{
diff_ratio(r1, r2)
}
\arguments{
\item{r1}{numeric vector; usually the sample signal-noise ratio.}

\item{r2}{numeric vector; usually the target signal-noise ratio.}
}
\description{
Compare the signal-noise ratios
}
