% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signal_noise_ratio.R
\name{extract_ratio}
\alias{extract_ratio}
\title{extract signal-noise ratio from GLM object.}
\usage{
extract_ratio(model_obj, absolute = TRUE)
}
\description{
extract signal-noise ratio from GLM object.
}

