% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_coefficients.R
\name{extract_ungrouped_data}
\alias{extract_ungrouped_data}
\title{Extract (ungrouped) data from the beta object}
\usage{
extract_ungrouped_data(data_model_obj)
}
\arguments{
\item{data_model_obj}{An object from 'generate_response_with_ratio'.}
}
\description{
Extract (ungrouped) data from the beta object
}
\details{
Not for external users.
}

