% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/response.R
\name{family2distFUN}
\alias{family2distFUN}
\title{Return the sampling distribution of the family object}
\usage{
family2distFUN(family, ...)
}
\description{
Return the sampling distribution of the family object
}

