% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/beta_coefficients.R
\name{generate_beta}
\alias{generate_beta}
\title{Generate predictor coefficients according to the signal-noise ratio}
\usage{
generate_beta(X, family, f = identity, target_ratio, max_iter = 100,
  tol = 0.1, curiosity = 1000, block_num = 20, lambda = 0)
}
\description{
Generate predictor coefficients according to the signal-noise ratio
}

