% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/beta_coefficients.R
\name{generate_beta_gaussian}
\alias{generate_beta_gaussian}
\title{Generate predictor coefficients according to the signal-noise ratio for Gaussian distribution}
\usage{
generate_beta_gaussian(X, target_ratio, ...)
}
\description{
Generate predictor coefficients according to the signal-noise ratio for Gaussian distribution
}

