% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_response.R
\name{generate_data}
\alias{generate_data}
\title{Simulate the response data given covariates, beta and DGP specification.}
\usage{
generate_data(X, beta, invLinkFUN, distFUN, f)
}
\description{
Simulate the response data given covariates, beta and DGP specification.
}

