% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_independent.R
\name{generate_features}
\alias{generate_features}
\title{Generate independent features - numerical or (ordinal) categorical}
\usage{
generate_features(n, p, distn = "gaussian", num_categories = 5, ...)
}
\arguments{
\item{n}{integer; number of data points to be simulated.}

\item{p}{integer; number of covariates.}

\item{distn}{One of 'gaussian', 'student-t', 'pareto', 'uniform', 'discrete_uniform', 'poisson', 'gamma'.}

\item{num_categories}{integer; number of categories for 'discrete_uniform'.}

\item{...}{Extra parameters for distributions other than 'gaussian' and 'uniform'.}
}
\description{
Generate independent features - numerical or (ordinal) categorical
}
\details{
See "?rnorm", "?rt", "?VGAM::rpareto", "?runif", "?rpois", "?rgamma"
for more details on the extra parameters. If 'discrete_uniform' is chosen, one
can specify the distribution for the categories through the parameter 'prob'.
}
