% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_covariates.R
\name{generate_independent_covariates}
\alias{generate_independent_covariates}
\title{Generate independent covariates - numerical or (ordinal) categorical}
\usage{
generate_independent_covariates(no_pts, no_covariates, type = "numerical",
  no_categories, ...)
}
\arguments{
\item{no_pts}{integer; number of data points to be simulated.}

\item{no_covariates}{integer; number of covariates.}

\item{type}{One of 'numerical' or 'categorical'.}

\item{no_categories}{integer; number of categories. Only applies when type = 'categorical'.}

\item{...}{See details.}
}
\description{
Generate independent covariates - numerical or (ordinal) categorical
}
\details{
If 'numerical' is chosen, one can specify the distribution through parameter 'distn'.
The choices are "gaussian", "student-t" or "pareto".
One should also provide the parameters when one specifies the distribution.
For more details, see "?rnorm", "?rt", "?VGAM::rpareto".
Note that for 'gaussian', standard normal is assumed if parameters are not given. \cr\cr
If 'categorical' is chosen, one can specify the distribution for the categories through the
parameter 'prob'.
}

