% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{generate_response}
\alias{generate_response}
\title{Simulate the response data given covariates, beta and DGP specification.}
\usage{
generate_response(X, beta, family, formula)
}
\arguments{
\item{X}{matrix; data containing the covariates}

\item{beta}{numeric vector; coefficients of the linear predictors.}

\item{family}{family object; see "?family" for more details.}

\item{f}{function; transformation for the covariates.}
}
\value{
a dataframe; the simulated data.
}
\description{
Simulate the response data given covariates, beta and DGP specification.
}
