% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/beta_coefficients.R
\name{generate_response_with_ratio}
\alias{generate_response_with_ratio}
\title{Generate predictor coefficients according to the signal-noise ratio (SNR)}
\usage{
generate_response_with_ratio(X, family, target_ratio, group = FALSE,
  f = identity, ...)
}
\arguments{
\item{X}{matrix; data containing the covariates}

\item{family}{family object; see "?family" for more details.}

\item{target_ratio}{numeric vector; the desired signal-noise ratio}

\item{group}{boolean; Should the data be grouped or ungrouped. See details.}

\item{f}{function; transformation for the covariates.}

\item{...}{extra parameters; see details.}
}
\description{
Generate predictor coefficients according to the signal-noise ratio (SNR)
}
\details{
For families other than 'gaussian', the function reweights the
samples such that the target signal-noise ratio is reached. This implies
the data records will be treated as grouped records. If 'group' is TRUE,
then the function returns the grouped data, otherwise, the function returns
the ungrouped data (through duplication). \cr\cr
If 'family' is gaussian, one can pass the 'sd' argument through `...`.
Otherwise, the function runs a stochastic search algorithm to search
for the predictors coefficients, and one can pass the 'max_iter' argument
(default to be 100) to `...`. \cr\cr
Other parameters are tol, curiosity and block_num. They are not recommended
to external users.
}

