% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/beta_coefficients.R
\name{generate_weights}
\alias{generate_weights}
\title{Optimise the weights matrix to match the target signal-noise ratio}
\usage{
generate_weights(X, family, f = identity, target_ratio, max_iter = 100,
  tol = 0.1, curiosity = 1000, block_num = 20)
}
\description{
Optimise the weights matrix to match the target signal-noise ratio
}

