% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmSimData.R
\docType{package}
\name{glmSimData}
\alias{glmSimData}
\alias{glmSimData-package}
\title{A package for simulating data}
\description{
This package simulates artifical data following a GLM model with dependent / independent, numerical / ordinal covariates.
}
\author{
Jackson Kwok
}

