% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signal_noise_ratio.R
\name{init_perf_fun}
\alias{init_perf_fun}
\title{Initialise objective function for optimisation}
\usage{
init_perf_fun(y, X, beta, family, item_label)
}
\description{
Initialise objective function for optimisation
}

