% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_noise_ratio.R
\name{init_signal_noise}
\alias{init_signal_noise}
\title{Initialise objective function for optimisation}
\usage{
init_signal_noise(my_data, family, tf = identity)
}
\description{
Initialise objective function for optimisation
}

