% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stochastic_search.R
\name{ls_loss}
\alias{ls_loss}
\title{Least-squared loss function}
\usage{
ls_loss(x, y)
}
\arguments{
\item{x}{numeric vector.}

\item{y}{numeric vector.}
}
\description{
Least-squared loss function
}

