% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{matrix_times_vector}
\alias{matrix_times_vector}
\title{Sum of matrix columns weighted by a vector}
\usage{
matrix_times_vector(m1, v1)
}
\arguments{
\item{m1}{a matrix}

\item{v1}{a vector}
}
\value{
a vector
}
\description{
Sum of matrix columns weighted by a vector
}
\examples{
\dontrun{
m0 <- matrix(1:9, nrow = 3); m0
v0 <- c(0.2, 0.2, 0.9); v0
matrix_times_vector( m0, v0 )
}
}
\keyword{internal}

