% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mimic_data.R
\name{mimic_binary_data}
\alias{mimic_binary_data}
\title{A function to mimic a dataset provided}
\usage{
mimic_binary_data(my_data0, yvar, beta_size, seed, mean = 0, sd = 0.5,
  load_balance = TRUE, load_range = c(-0.1, 0.1))
}
\arguments{
\item{my_data0}{dataframe; the data to be mimic.}

\item{yvar}{character; name of the response variable.}

\item{beta_size}{integer; number of non-zero beta.}

\item{seed}{integer; seed for random number generator. Not to be used with load_balance.}

\item{mean}{numeric; beta ~ N(mean, sd)}

\item{sd}{numeric; beta ~ N(mean, sd)}

\item{load_balance}{logical; whether the simulated data should contain roughly equal number of positive cases as the original data.}

\item{load_range}{numeric vector; desirable range of proportions of positive cases.}
}
\description{
A function to mimic a dataset provided
}

