% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mimic_data.R
\name{mimic_binary_data_core}
\alias{mimic_binary_data_core}
\title{Core function to mimic a dataset provided}
\usage{
mimic_binary_data_core(my_data0, yvar, beta_size, seed, mean = 0, sd = 0.5)
}
\description{
Core function to mimic a dataset provided
}

