% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/beta_coefficients.R
\name{modify_beta}
\alias{modify_beta}
\title{Reduce the weights such that it rounds to multiples of 1000.}
\usage{
modify_beta(beta, N)
}
\description{
Reduce the weights such that it rounds to multiples of 1000.
}

