% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signal_noise_ratio.R
\name{random_labels}
\alias{random_labels}
\title{Generate some random labels}
\usage{
random_labels(num_label, labels = c("S", "M", "L"))
}
\arguments{
\item{num_label}{integer; number of labels desired.}

\item{labels}{vector; labels to be drawn from.}
}
\value{
character vector; the labels.
}
\description{
Generate some random labels
}

