% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_noise_ratio.R
\name{random_signals}
\alias{random_signals}
\title{Generate some random signals based on labels}
\usage{
random_signals(labels, labels_table)
}
\arguments{
\item{labels}{vector; the desired labels}

\item{labels_table}{dataframe; a table mapping labels to intervals.
See `create_labels_table()` for an example.}
}
\description{
Given a label, the function draws a number from the interval
corresponding to the label.
}
\examples{
labels_table <- create_labels_table()
random_signals(c(rep("S", 4), rep("M", 3)), labels_table)
}
