% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_coefficients.R
\name{reduce_data_weights}
\alias{reduce_data_weights}
\title{Reduce the weights such that it rounds to 'nice' numbers.}
\usage{
reduce_data_weights(data_model_obj, N)
}
\arguments{
\item{data_model_obj}{An object from 'generate_response_with_ratio'.}

\item{N}{Number of weights to be trimmed.}
}
\description{
Reduce the weights such that it rounds to 'nice' numbers.
}
\details{
Not for external users.
}

