% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_coefficients.R
\name{response_with_ratio}
\alias{response_with_ratio}
\title{[Core] Optimise the weights matrix to match the target signal-noise ratio}
\usage{
response_with_ratio(X, family, f = identity, target_ratio, max_iter = 100,
  tol = 0.1, curiosity = 1000, block_num = 20)
}
\description{
[Core] Optimise the weights matrix to match the target signal-noise ratio
}

