% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_coefficients.R
\name{response_with_ratio_gaussian}
\alias{response_with_ratio_gaussian}
\title{[Core] Generate predictor coefficients according to the signal-noise ratio for Gaussian distribution}
\usage{
response_with_ratio_gaussian(X, target_ratio, ...)
}
\description{
[Core] Generate predictor coefficients according to the signal-noise ratio for Gaussian distribution
}

