% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_DGP.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data}
\usage{
simulate_data(no_pts = 5000, no_covariates, beta_spec, family = gaussian(),
  f = identity, independent_covariates = TRUE, type = "numerical", ...)
}
\arguments{
\item{no_pts}{integer number of data points to be simulated.}

\item{no_covariates}{number of covariates to be simulated.}

\item{beta_spec}{vector of "effect size" of the covariates, can be one of "XL", "L", "M", "S", "XS".
See details for more information of the encoding.}

\item{family}{GLM family.}

\item{f}{the function relationship between the response and the explanatory variables.}

\item{independent_covariates}{whether the covariates are independent.}

\item{type}{the type of covariates, either "numerical" or "categorical".}

\item{...}{extra parameters corresponding to different choices of independent_covariates.
See "?generate_independent_covariates" or "?generate_dependent_covariates" for more details.}

\item{no_covariates}{integer; excluding constant.}
}
\description{
Simulate data
}

