% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_search.R
\name{stochastic_search}
\alias{stochastic_search}
\title{Stochastic Search}
\usage{
stochastic_search(dim_param, perf_fun, loss_fun = ls_loss, target_perf,
  max_iter = 100, tol = 0, curiosity = 1, block_num, block_size,
  lambda = 0, reg_fun = zero_reg, param)
}
\description{
Stochastic Search
}

