% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{trim_number}
\alias{trim_number}
\title{Trim the first s significant digits.}
\usage{
trim_number(n, s = 2)
}
\arguments{
\item{n}{integer; number to be trimmed.}

\item{s}{integer; number of significant to be trimmed.}
}
\description{
Trim the first s significant digits.
}
\examples{
trim_number(1337, 1)  # gives 337
trim_number(1337, 2)  # gives 37
}

