% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signal_noise_ratio.R
\name{vec_find_signal_label}
\alias{vec_find_signal_label}
\title{Categorise signal-noise ratio (vectorised)}
\usage{
vec_find_signal_label(signal_noise_ratio)
}
\arguments{
\item{signal_noise_ratio}{numeric vector;}
}
\value{
character vector; the labels.
}
\description{
Categorise signal-noise ratio (vectorised)
}

