% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_noise_ratio.R
\name{extract_ratio}
\alias{extract_ratio}
\title{extract signal-noise ratio from GLM object.}
\usage{
extract_ratio(model_obj, absolute = TRUE)
}
\arguments{
\item{model_obj}{'glm' object.}

\item{absolute}{boolean; if FALSE, return the signed ratio.}
}
\description{
extract signal-noise ratio from GLM object.
}
\details{
The signal-noise ratio must be non-negative, but sometimes the
signs are needed for analysis.
}
