% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_beta.R
\name{find_signal_label}
\alias{find_signal_label}
\title{Categorise signal-noise ratio
<--XS-- 0.22 --S-- 0.61 --M-- 1.65 --L-- 4.48 --XL-->}
\usage{
find_signal_label(signal_noise_ratio)
}
\description{
Categorise signal-noise ratio
<--XS-- 0.22 --S-- 0.61 --M-- 1.65 --L-- 4.48 --XL-->
}

