% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_beta.R
\name{generate_beta}
\alias{generate_beta}
\title{A function to generate beta coefficients with desired effect size.}
\usage{
generate_beta(X, beta_spec, family, invLinkFUN, distFUN, f, iter = 100)
}
\arguments{
\item{beta_spec}{character vector; element can be one of
'XL', 'L', 'M', 'S', 'XS', suggesting the size of the effect.}
}
\value{
a vector, the beta coefficients
}
\description{
A function to generate beta coefficients with desired effect size.
}
\details{
The effect size of the betas are determined by signal-to-noise ratio,
To find that we need to simulate some data and use the empirical ratio as guide.
}

