% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/beta_coefficients.R
\name{generate_beta}
\alias{generate_beta}
\title{Generate predictor coefficients according to the signal-noise ratio}
\usage{
generate_beta(X, family, f = identity, target_ratio, group = TRUE, ...)
}
\description{
Generate predictor coefficients according to the signal-noise ratio
}

