% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/response.R
\name{generate_response}
\alias{generate_response}
\title{Simulate the response data given covariates, beta and DGP specification.}
\usage{
generate_response(X, beta, invLinkFUN, distFUN, f)
}
\description{
Simulate the response data given covariates, beta and DGP specification.
}

