% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_coefficients.R
\name{reduce_data_weights}
\alias{reduce_data_weights}
\title{Reduce the weights such that it rounds to 'nice' numbers.}
\usage{
reduce_data_weights(data_model_obj, N, by = 1, choose = "random", family)
}
\arguments{
\item{data_model_obj}{An object from 'generate_response_with_ratio'.}

\item{N}{Number of weights to be trimmed.}

\item{by}{Integer; desired decrement for each weight.}

\item{choose}{One of 'random' and 'top'.
If 'random', the function reduces a random set of N weights.
If 'top', the function reduces the heaviest N weights.}

\item{family}{GLM family, see "?family" for details.}
}
\description{
Reduce the weights such that it rounds to 'nice' numbers.
}
\details{
Not for external users.
}

